Match LV2 specification instantiate prototype. Added (in-progress) OSC Metronome plugin.
git-svn-id: http://svn.drobilla.net/lad/slv2@605 a436a847-0d15-0410-975c-d299462d15a1

+++ slv2/plugininstance.h
@@ -73,7 +73,7 @@ typedef struct _Instance {
  */
 SLV2Instance
 slv2_plugin_instantiate(SLV2Plugin               plugin,
-                        uint32_t                 sample_rate,
+                        double                   sample_rate,
                         const LV2_Host_Feature** host_features);
