+++ src/plugininstance.c
@@ -32,7 +32,7 @@
 
 SLV2Instance
 slv2_plugin_instantiate(SLV2Plugin               plugin,
-                        uint32_t                 sample_rate,
+                        double                   sample_rate,
                         const LV2_Host_Feature** host_features)
 {
 	struct _Instance* result = NULL;
